+++ sample-core/src/commonMain/kotlin/com/mobilejazz/kmmsample/core/feature/hackerposts/HackerNewsPostsProvider.kt
@@ -60,7 +60,7 @@ class HackerNewsPostsDefaultModule(
     val hackerPostsIdsNetworkDataSource = GetNetworkDataSource(
       networkConfiguration.baseUrl,
       networkConfiguration.httpClient,
-      ListSerializer(Int.serializer()),
+      ListSerializer(Long.serializer()),
       networkConfiguration.json
     )
