+++ sample-core/src/commonMain/kotlin/com/mobilejazz/kmmsample/core/feature/hackerposts/domain/model/HackerNewsPostsIds.kt
@@ -1,3 +1,3 @@
 package com.mobilejazz.kmmsample.core.feature.hackerposts.domain.model
 
-data class HackerNewsPostsIds(val listIds: List<Int>)
+data class HackerNewsPostsIds(val listIds: List<Long>)
